backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.0.1
django-storages==1.14.2
boto3==1.28.62
cryptography==41.0.4
celery==5.3.4
redis==5.0.1
django-redis==5.2.0
drf-yasg==1.21.5
django-filter==23.3
django-allauth==0.57.0
stripe==7.0.0
#python-mpesa==0.5.0
django-ratelimit==3.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.1.0
pytest==7.4.2
pytest-django==4.7.0
factory-boy==3.2.1
Faker==19.9.0
coverage==7.3.2
